+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Python 3.2 and 3.4 support
 * Improved test suite - 100% coverage + flake8
+* Multiple database support - `TEST_DATABASE_URL` will create `settings.DATABASES['test']`
 
 ## 1.2